requirements.txt
python-dotenv~=1.1.0
deepl~=1.21.1
langdetect~=1.0.9
google-cloud-translate
pip~=24.0
protobuf~=5.29.4
six~=1.17.0
requests~=2.32.3
idna~=3.10
certifi~=2025.4.26
charset-normalizer~=3.4.1
urllib3~=2.4.0